Avoid purging too much data from var on "unprovision --rid ..."
Don't purge the scheduler status, not the service cgroup.
Purge resource var dirs only for the selected resources.

One visible effect of this bug was the volumes going "down" after unprovisioning
another resource.

+++ lib/svc.py
@@ -2502,7 +2502,10 @@ class BaseSvc(Crypt, ExtConfigMixin):
             pass
 
     def purge_var_d(self):
+        scoped = self.command_is_scoped()
         for res in self.get_resources():
+            if scoped and res.skip:
+                continue
             res.purge_var_d()
 
     def delete_service_conf(self):
@@ -4394,8 +4397,9 @@ class Svc(BaseSvc):
     def unprovision(self):
         self.sub_set_action("disk.scsireserv", "stop", xtags=set(["zone", "docker", "podman"]))
         self.sub_set_action(STOP_GROUPS, "unprovision", xtags=set(["zone", "docker", "podman"]))
-        self.pg_remove()
-        self.delete_service_sched()
+        if not self.command_is_scoped():
+            self.pg_remove()
+            self.delete_service_sched()
         self.purge_var_d()
 
     def provision(self):
